% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{clean_numeric_column}
\alias{clean_numeric_column}
\title{Clean numeric column by removing special characters}
\usage{
clean_numeric_column(x)
}
\arguments{
\item{x}{Vector to clean}
}
\value{
Numeric vector
}
\description{
Clean numeric column by removing special characters
}
