% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_filter_utils.R
\name{create_custom_filter_example}
\alias{create_custom_filter_example}
\title{Create Custom Filter Example}
\usage{
create_custom_filter_example(function_name, os = "ios")
}
\arguments{
\item{function_name}{Character. Name of the function}

\item{os}{Character. Example OS to use}
}
\value{
Character string with example code
}
\description{
Generates example code for using custom filters with a function
}
\keyword{internal}
