% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_id_optimization.R
\name{.onLoad}
\alias{.onLoad}
\title{Package Load Hook for ID Optimization}
\usage{
.onLoad(libname, pkgname)
}
\description{
Sets up ID caching and optimization on package load
}
\keyword{internal}
