% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_helpers.R
\name{format_arpu}
\alias{format_arpu}
\title{Format ARPU (Average Revenue Per User)}
\usage{
format_arpu(val, digits = 2)
}
\arguments{
\item{val}{Numeric ARPU value}

\item{digits}{Number of decimal places (default: 2)}
}
\value{
Formatted ARPU string
}
\description{
Format ARPU (Average Revenue Per User)
}
\examples{
format_arpu(5.234)    # "$5.23"
format_arpu(0.99)     # "$0.99"
}
