% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_helpers.R
\name{format_currency}
\alias{format_currency}
\title{Format currency values with appropriate suffixes}
\usage{
format_currency(val, digits = 2)
}
\arguments{
\item{val}{Numeric value to format as currency}

\item{digits}{Number of digits after decimal for millions/billions (default: 2)}
}
\value{
Formatted currency string
}
\description{
Format currency values with appropriate suffixes
}
\examples{
format_currency(1234567)      # "$1.23M"
format_currency(1234567890)   # "$1.23B"
format_currency(123)          # "$123"
}
