% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_helpers.R
\name{format_downloads}
\alias{format_downloads}
\title{Format download counts with appropriate suffixes}
\usage{
format_downloads(val)
}
\arguments{
\item{val}{Numeric value to format as downloads}
}
\value{
Formatted download count string
}
\description{
Format download counts with appropriate suffixes
}
\examples{
format_downloads(1234567)      # "1.2M"
format_downloads(1234567890)   # "1.2B"
}
