% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_helpers.R
\name{format_percent}
\alias{format_percent}
\title{Format percentages}
\usage{
format_percent(val, digits = 1)
}
\arguments{
\item{val}{Numeric value to format as percentage (0-100 scale)}

\item{digits}{Number of decimal places (default: 1)}
}
\value{
Formatted percentage string
}
\description{
Format percentages
}
\examples{
format_percent(23.456)    # "23.5\%"
format_percent(0.234, digits = 2)    # "0.23\%"
}
