% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_helpers.R
\name{format_retention}
\alias{format_retention}
\title{Format retention rates}
\usage{
format_retention(val, digits = 1)
}
\arguments{
\item{val}{Numeric value as decimal (0-1 scale)}

\item{digits}{Number of decimal places (default: 1)}
}
\value{
Formatted retention percentage string
}
\description{
Format retention rates
}
\examples{
format_retention(0.234)    # "23.4\%"
format_retention(0.85)     # "85.0\%"
}
