% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_helpers.R
\name{format_users}
\alias{format_users}
\title{Format user counts (DAU/MAU/WAU)}
\usage{
format_users(val)
}
\arguments{
\item{val}{Numeric value to format as user count}
}
\value{
Formatted user count string
}
\description{
Format user counts (DAU/MAU/WAU)
}
\examples{
format_users(1234567)      # "1.2M"
format_users(1234)         # "1.2K"
}
