% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_helpers.R
\name{get_column_spec}
\alias{get_column_spec}
\title{Get column specification for common Sensor Tower metrics}
\usage{
get_column_spec(metric_type = NULL, time_period = NULL, region = NULL)
}
\arguments{
\item{metric_type}{Type of metric (revenue, downloads, retention, etc.)}

\item{time_period}{Time period (30d, 180d, alltime, etc.)}

\item{region}{Region (us, ww)}
}
\value{
List of column specifications
}
\description{
Get column specification for common Sensor Tower metrics
}
