% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensor_tower_api.R
\name{get_unified_app_id}
\alias{get_unified_app_id}
\title{Get Unified App ID from Publisher}
\usage{
get_unified_app_id(auth_token, publisher_id, app_name)
}
\arguments{
\item{auth_token}{Sensor Tower API authentication token}

\item{publisher_id}{Publisher ID from Sensor Tower}

\item{app_name}{Name of the app to search for}
}
\value{
The unified app ID for the specified app
}
\description{
Get Unified App ID from Publisher
}
