% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_id_cache.R
\docType{data}
\name{id_cache}
\alias{id_cache}
\alias{.sensortowerR_env}
\title{Intelligent App ID Cache and Resolution System}
\format{
An object of class \code{environment} of length 0.
}
\usage{
.sensortowerR_env
}
\description{
Internal functions to manage a persistent cache of app ID mappings,
reducing API calls and improving performance.
}
\keyword{internal}
