% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_sales_report.R
\name{lookup_category_names}
\alias{lookup_category_names}
\title{Helper function to look up category names}
\usage{
lookup_category_names(category_ids, platform = "ios")
}
\arguments{
\item{category_ids}{Character vector of category IDs}

\item{platform}{Character string. "ios" or "android"}
}
\value{
Character vector of category names
}
\description{
Helper function to look up category names
}
