% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{NUMERIC_METRIC_PATTERNS}
\alias{NUMERIC_METRIC_PATTERNS}
\title{Numeric Metric Patterns}
\format{
An object of class \code{character} of length 29.
}
\usage{
NUMERIC_METRIC_PATTERNS
}
\description{
Patterns to identify columns that should be converted to numeric.
}
\keyword{internal}
