% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_app_details.R
\name{process_app_details_response}
\alias{process_app_details_response}
\title{Process App Details API Response}
\usage{
process_app_details_response(resp, os)
}
\arguments{
\item{resp}{Response object from httr2}

\item{os}{Operating system}
}
\value{
A processed tibble with app details
}
\description{
Internal function to process and enrich app details API responses.
}
\keyword{internal}
