% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_category_rankings.R
\name{process_ranking_response}
\alias{process_ranking_response}
\title{Process Category Ranking API Response}
\usage{
process_ranking_response(resp, os, category, country, chart_type, date, limit)
}
\arguments{
\item{resp}{Response object from httr2}

\item{os}{Operating system}

\item{category}{Category ID used in request}

\item{country}{Country code used in request}

\item{chart_type}{Chart type used in request}

\item{date}{Date used in request (may be NULL)}

\item{limit}{Number of results requested}
}
\value{
A processed tibble with ranking data
}
\description{
Internal function to process and enrich category ranking API responses.
}
\keyword{internal}
