% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{select_robust}
\alias{select_robust}
\title{Safe column selection with automatic fallbacks}
\usage{
select_robust(data, ...)
}
\arguments{
\item{data}{Data frame}

\item{...}{Column specifications (can be character vectors)}
}
\value{
Data frame with selected columns
}
\description{
Safe column selection with automatic fallbacks
}
