% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{sensortowerR-defunct}
\alias{sensortowerR-defunct}
\alias{st_sales_report}
\alias{st_unified_sales_report}
\alias{st_batch_metrics}
\alias{st_smart_metrics}
\alias{st_top_charts}
\alias{st_top_publishers}
\alias{st_category_rankings}
\alias{st_app_info}
\alias{st_app_lookup}
\alias{st_app_details}
\alias{st_filter_by_date}
\alias{st_filter_by_genre}
\alias{st_filter_by_monetization}
\alias{st_filter_by_publisher}
\alias{st_filter_by_sdk}
\alias{st_custom_fields_filter}
\alias{st_custom_fields_filter_by_id}
\alias{st_combine_filters}
\alias{st_create_simple_filter}
\alias{st_compare_filter_results}
\alias{st_generate_example_filter_ids}
\title{Defunct Functions In sensortowerR}
\usage{
st_sales_report(...)

st_unified_sales_report(...)

st_batch_metrics(...)

st_smart_metrics(...)

st_top_charts(...)

st_top_publishers(...)

st_category_rankings(...)

st_app_info(...)

st_app_lookup(...)

st_app_details(...)

st_filter_by_date(...)

st_filter_by_genre(...)

st_filter_by_monetization(...)

st_filter_by_publisher(...)

st_filter_by_sdk(...)

st_custom_fields_filter(...)

st_custom_fields_filter_by_id(...)

st_combine_filters(...)

st_create_simple_filter(...)

st_compare_filter_results(...)

st_generate_example_filter_ids(...)
}
\description{
Legacy entrypoints retained only as defunct stubs for the v1.0.0 transition.
}
\keyword{internal}
