% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{st_clear_app_cache}
\alias{st_clear_app_cache}
\title{Clear App Name Cache}
\usage{
st_clear_app_cache()
}
\description{
Clears the internal cache of app name lookups. Useful for testing or when
you want to refresh app name data.
}
