% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_smart_metrics.R
\name{st_clear_id_cache}
\alias{st_clear_id_cache}
\title{Clear App ID Cache}
\usage{
st_clear_id_cache(disk = TRUE)
}
\arguments{
\item{disk}{Logical. Also remove the on-disk cache file (default TRUE).}
}
\description{
Clears the in-memory and on-disk cache of app ID mappings
}
