% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{st_filter_read}
\alias{st_filter_read}
\title{Read a server filter definition}
\usage{
st_filter_read(filter, auth_token = NULL)
}
\arguments{
\item{filter}{A created st_filter or existing server ID.}

\item{auth_token}{API token; defaults to SENSORTOWER_AUTH_TOKEN.}
}
\value{
A tibble of custom field predicates with filter_id provenance.
}
\description{
Read a server filter definition
}
