% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_entities.R
\name{search_entities}
\alias{search_entities}
\title{Search for apps or publishers in Sensor Tower}
\usage{
search_entities(
  term,
  os = "unified",
  entity_type = "app",
  limit = 100,
  auth_token = Sys.getenv("sensortower_auth")
)
}
\arguments{
\item{term}{Search term (min 3 Latin chars or 2 non-Latin chars)}

\item{os}{Platform (unified)}

\item{entity_type}{Type of entity to search for (app or publisher)}

\item{limit}{Number of results to return (max 250)}

\item{auth_token}{Authentication token}
}
\value{
List containing search results
}
\description{
Search for apps or publishers in Sensor Tower
}
\examples{
\dontrun{
search_entities(term = "Spotify")
}
}
