% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_entities.R
\name{search_entities}
\alias{search_entities}
\title{Search for apps or publishers}
\usage{
search_entities(
  os = "unified",
  entity_type = "app",
  term,
  limit = 100,
  auth_token = Sys.getenv("SENSORTOWER_AUTH")
)
}
\arguments{
\item{os}{Platform (unified)}

\item{entity_type}{Type of entity to search for (app or publisher)}

\item{term}{Search term (min 3 Latin chars or 2 non-Latin chars)}

\item{limit}{Number of results to return (max 250)}

\item{auth_token}{Authentication token}
}
\value{
List containing search results
}
\description{
Search for apps or publishers
}
\examples{
\dontrun{
search_entities(os = "unified", entity_type = "app", term = "Lyft", limit = 100)
}
}
